Move code under Core namespace
