+++ net/src/Sails.Remoting.Abstractions/Core/IRemoting.cs
@@ -5,7 +5,7 @@ using Substrate.Gear.Api.Generated.Model.gprimitives;
 using GasUnit = Substrate.NetApi.Model.Types.Primitive.U64;
 using ValueUnit = Substrate.NetApi.Model.Types.Primitive.U128;
 
-namespace Sails.Remoting.Abstractions;
+namespace Sails.Remoting.Abstractions.Core;
 
 public interface IRemoting
 {
