+++ net/src/Sails.Remoting/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,11 +1,12 @@
 ﻿using EnsureThat;
 using Microsoft.Extensions.DependencyInjection;
-using Sails.Remoting.Abstractions;
+using Sails.Remoting.Abstractions.Core;
+using Sails.Remoting.Core;
 using Sails.Remoting.Options;
 
 namespace Sails.Remoting.DependencyInjection;
 
-public static class IServiceCollectionExtensions
+public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddRemotingViaNodeClient(
         this IServiceCollection services,
